Countries are now listed as participating
Some countries only vote in particular events, this has been
added into the JSON as well as a new structure.

Country data is now a map with "participating" and "votes" as
keys. A KeyError check is required in the vote extraction loop
to check for non-participating voters.

JSON is now indented when written to the file