Script now dumps a map of country name -> alpha3
This is dumped to countries.json and features every country
mentioned in every event.

A major bug has been fixed where no-competing voters were not
added to participants of a particular event. This was due to
attempting to address the countries dictionary with the
contestant country name rather than the voter's when checking for
a KeyError.